kafka_consumer/requirements.txt
boto3
kafka-python
aws-glue-schema-registry
